requirements.txt
aioitertools==0.11.0
annotated-types==0.6.0
anyio==4.2.0
ase==3.22.1
-e git+https://github.com/tsaie79/atomate@35b9e2dd9022406ea5aa87a503805a03b63406a3#egg=atomate
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.30
botocore==1.34.30
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.1
custodian==2023.10.9
cycler==0.12.1
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
FireWorks==2.0.3
Flask==3.0.1
flask-paginate==2023.10.24
fonttools==4.47.2
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
latexcodec==2.0.1
lazy_loader==0.3
maggma==0.61.1
MarkupSafe==2.1.4
matplotlib==3.8.2
mongogrant==0.3.3
mongomock==4.1.2
monty==2024.1.26
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.3
orjson==3.9.12
packaging==23.2
palettable==3.3.3
pandas==2.2.0
paramiko==3.4.0
pillow==10.2.0
plotly==5.18.0
prettytable==3.9.0
psutil==5.9.8
pybtex==0.24.0
pycparser==2.21
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pydash==7.0.7
-e git+https://github.com/tsaie79/pymatgen.git@737b893386478d8b911f16e7cc347462477d0b4f#egg=pymatgen
pymatgen-analysis-defects==2024.1.12
pymatgen-analysis-diffusion==2023.8.15
pymongo==4.6.1
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scikit-image==0.22.0
scipy==1.12.0
sentinels==1.0.0
sentry-sdk==1.39.2
six==1.16.0
sniffio==1.3.0
spglib==2.3.0
sshtunnel==0.4.0
starlette==0.35.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tifffile==2024.1.30
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uncertainties==3.1.7
urllib3==1.26.18
uvicorn==0.27.0.post1
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
